Conflicts:
	src/voglcommon/vogl_gl_replayer.cpp
	src/voglcommon/vogl_gl_utils.cpp
	src/voglcommon/vogl_gl_utils.h
	src/vogltrace/vogl_intercept.cpp

+++ src/vogleditor/vogleditor_qtextureviewer.cpp
@@ -28,18 +28,18 @@
 #include "vogl_buffer_stream.h"
 
 QTextureViewer::QTextureViewer(QWidget *parent) :
-   QWidget(parent),
-   m_channelSelection(VOGL_CSO_RGBA),
-   m_zoomFactor(1),
-   m_bInvert(false),
-   m_pKtxTexture(NULL),
-   m_baseMipLevel(0),
-   m_maxMipLevel(0),
-   m_arrayIndex(0)
+    QWidget(parent),
+    m_channelSelection(VOGL_CSO_RGBA),
+    m_zoomFactor(1),
+    m_bInvert(false),
+    m_pKtxTexture(NULL),
+    m_baseMipLevel(0),
+    m_maxMipLevel(0),
+    m_arrayIndex(0)
 {
-   m_background = QBrush(QColor(0, 0, 0));
-   m_outlinePen = QPen(Qt::black);
-   m_outlinePen.setWidth(1);
+    m_background = QBrush(QColor(0, 0, 0));
+    m_outlinePen = QPen(Qt::black);
+    m_outlinePen.setWidth(1);
 }
 
 void QTextureViewer::setTexture(const vogl::ktx_texture* pTexture, uint baseMipLevel, uint maxMipLevel)
@@ -75,12 +75,12 @@ void QTextureViewer::paintEvent(QPaintEvent *event)
 
     if (m_draw_enabled == false)
     {
-       // clear the viewer
-       painter.fillRect(event->rect(), QWidget::palette().color(QWidget::backgroundRole()));
+        // clear the viewer
+        painter.fillRect(event->rect(), QWidget::palette().color(QWidget::backgroundRole()));
     }
     else
     {
-       paint(&painter, event);
+        paint(&painter, event);
     }
     painter.end();
 }
@@ -232,13 +232,13 @@ void QTextureViewer::paint(QPainter *painter, QPaintEvent *event)
 
         if (m_pixmaps.contains(mip))
         {
-            uint left = 0;
-            uint top = 0;
+            int left = 0;
+            int top = 0;
             if (m_bInvert)
             {
                 // invert
                 painter->scale(1,-1);
-                top = -drawHeight;
+                top = -(int)drawHeight;
             }
 
             painter->drawPixmap(left, top, drawWidth, drawHeight, m_pixmaps[mip]);
